MDEV-26238: Remove inconsistent behaviour of --default-* options in my_print_defaults
Analysis: --defaults* option is recognized anywhere in the commandline
instead of only at the beginning because handle_options() recognizes
options in any order.
Fix: use get_defaults_options() which recognizes --defaults* options only at
the beginning. After this is done, we only want to recognize other options
given in any order which can be done using handle_options(). But
handle_options() returns error and exits if an option is not found.
We don't want to exit and give error because --defaults* options are still
valid. So make a new commandline which doesn't have --defaults*
option and pass it to handle_options(). Also, removed -e, -g and -c because
only my_print_defaults supports them.