MDEV-26467: Avoid futile spin loops
Typically, index_lock, block_lock and fil_space_t::latch
will be held for a longer time than the spinloop in latch acquisition
would be waiting for. Let us avoid spinloops for those as well as
dict_sys.latch, which could be held in exclusive mode for a longer time
(while loading metadata into the buffer pool and the dictionary cache).

We introduce srw_spin_lock and srw_spin_mutex to allow users of
srw_lock or srw_mutex to opt in for the spin loop.
On Microsoft Windows, a spin loop variant will not be available;
srw_mutex and srw_lock will simply wrap SRWLOCK.
That is, the parameters innodb_sync_spin_loops and
innodb_spin_wait_delay will cease to have any effect on
Microsoft Windows.